+++ service/database/getUserProfile.go
@@ -7,7 +7,7 @@ import (
 )
 
 func (db *appdbimpl) GetUserProfile(u User) (User, error) {
-	var id, profilepic uint64
+	var id, profilepic int
 	var name, followers, banned, photos string
 	rows, err := db.c.Query(`select id, name,profilepic,followers,banned, photos from users where id=?`, u.ID) //Here followers will be a string, then casted to string array
